Move lockfile model into versioned package
